refactor: NxHailo.Hailo -> NxHailo
